+++ NEWS.md
@@ -2,7 +2,7 @@
 
 # development version
 
-- Add Yoon 2023 vancomycin ([doi:10.3389/fphar.2023.1252757](https://doi.org/10.3389/fphar.2023.1252757)) -- hospitalised Korean adults and children receiving intravenous vancomycin.
+- Add Verrest 2023 paromomycin and miltefosine ([doi:10.1093/jac/dkad286](https://doi.org/10.1093/jac/dkad286)) -- Eastern African children and adults with visceral leishmaniasis.
 
 - Add Ousey 2026 plozasiran ([doi:10.1002/jcph.70190](https://doi.org/10.1002/jcph.70190)) -- adults with familial chylomicronemia syndrome.